+++ examples/chat/src/main.rs
@@ -3,7 +3,6 @@ use std::net;
 use std::str::FromStr;
 
 use actix::prelude::*;
-use futures_util::stream::StreamExt;
 use tokio::net::{TcpListener, TcpStream};
 use tokio_util::codec::FramedRead;
 
@@ -56,7 +55,15 @@ async fn main() {
 
     // Create server listener
     let addr = net::SocketAddr::from_str("127.0.0.1:12345").unwrap();
-    let listener = Box::new(TcpListener::bind(&addr).await.unwrap());
+    let listener = TcpListener::bind(&addr).await.unwrap();
+
+    // create a stream with the help of stream macro.
+    let stream = async_stream::stream! {
+        // accept incoming tcp stream and remote addr and return them as a stream
+        while let Ok((st, addr)) = listener.accept().await {
+            yield TcpConnect(st, addr);
+        }
+    };
 
     // Our chat server `Server` is an actor, first we need to start it
     // and then add stream on incoming tcp connections to it.
@@ -64,11 +71,7 @@ async fn main() {
     // items So to be able to handle this events `Server` actor has to implement
     // stream handler `StreamHandler<(TcpStream, net::SocketAddr), io::Error>`
     Server::create(move |ctx| {
-        ctx.add_message_stream(Box::leak(listener).incoming().map(|st| {
-            let st = st.unwrap();
-            let addr = st.peer_addr().unwrap();
-            TcpConnect(st, addr)
-        }));
+        ctx.add_message_stream(stream);
         Server { chat: server }
     });