+++ src/address/message.rs
@@ -7,7 +7,7 @@ use futures_channel::oneshot;
 
 use pin_project::pin_project;
 
-use crate::clock::Delay;
+use crate::clock::Sleep;
 use crate::handler::{Handler, Message};
 
 use super::channel::{AddressSender, Sender};
@@ -25,7 +25,8 @@ where
 {
     rx: Option<oneshot::Receiver<M::Result>>,
     info: Option<(AddressSender<A>, M)>,
-    timeout: Option<Delay>,
+    #[pin]
+    timeout: Option<Sleep>,
 }
 
 impl<A, M> Request<A, M>
@@ -52,16 +53,16 @@ where
 
     /// Set message delivery timeout
     pub fn timeout(mut self, dur: Duration) -> Self {
-        self.timeout = Some(tokio::time::delay_for(dur));
+        self.timeout = Some(tokio::time::sleep(dur));
         self
     }
 
     fn poll_timeout(
-        &mut self,
+        self: Pin<&mut Self>,
         cx: &mut task::Context<'_>,
     ) -> Poll<Result<M::Result, MailboxError>> {
-        if let Some(ref mut timeout) = self.timeout {
-            match Pin::new(timeout).poll(cx) {
+        if let Some(timeout) = self.project().timeout.as_pin_mut() {
+            match timeout.poll(cx) {
                 Poll::Ready(()) => Poll::Ready(Err(MailboxError::Timeout)),
                 Poll::Pending => Poll::Pending,
             }
@@ -118,7 +119,8 @@ where
 {
     rx: Option<oneshot::Receiver<M::Result>>,
     info: Option<(Box<dyn Sender<M>>, M)>,
-    timeout: Option<Delay>,
+    #[pin]
+    timeout: Option<Sleep>,
 }
 
 impl<M> RecipientRequest<M>
@@ -139,16 +141,16 @@ where
 
     /// Set message delivery timeout
     pub fn timeout(mut self, dur: Duration) -> Self {
-        self.timeout = Some(tokio::time::delay_for(dur));
+        self.timeout = Some(tokio::time::sleep(dur));
         self
     }
 
     fn poll_timeout(
-        &mut self,
+        self: Pin<&mut Self>,
         cx: &mut task::Context<'_>,
     ) -> Poll<Result<M::Result, MailboxError>> {
-        if let Some(ref mut timeout) = self.timeout {
-            match Pin::new(timeout).poll(cx) {
+        if let Some(timeout) = self.project().timeout.as_pin_mut() {
+            match timeout.poll(cx) {
                 Poll::Ready(()) => Poll::Ready(Err(MailboxError::Timeout)),
                 Poll::Pending => Poll::Pending,
             }