+++ tests/test_connected.rs
@@ -1,6 +1,6 @@
 use actix::prelude::*;
 use std::time::Duration;
-use tokio::time::delay_for;
+use tokio::time::sleep;
 
 struct MyActor;
 
@@ -24,7 +24,7 @@ impl Actor for MyActor {
 async fn test_connected() {
     Arbiter::spawn(async move {
         let addr = MyActor::start(MyActor);
-        delay_for(Duration::from_millis(350)).await;
+        sleep(Duration::from_millis(350)).await;
         drop(addr);
     });
 }